refactor: move templates to sections
