+++ packages/app/src/modules/options/gaugeConfig.js
@@ -7,10 +7,10 @@ import AxisRange from '../../components/VisualizationOptions/Options/AxisRange'
 import SeriesAxisLabels from '../../components/VisualizationOptions/Options/SeriesAxisLabels'
 import getSeriesTab from './tabs/series'
 import getAxesTab from './tabs/axes'
-import getVerticalAxisTemplate from './templates/verticalAxis'
+import getVerticalAxisTemplate from './sections/templates/verticalAxis'
 import getDataTab from './tabs/data'
 import getAdvancedSection from './sections/advanced'
-import getLinesTemplate from './templates/lines'
+import getLinesTemplate from './sections/templates/lines'
 import getLegendTab from './tabs/legend'
 import getStyleTab from './tabs/style'
 import getTitlesSection from './sections/titles'
