+++ packages/app/src/modules/options/pivotTableConfig.js
@@ -29,10 +29,10 @@ import ShowHierarchy from '../../components/VisualizationOptions/Options/ShowHie
 import CompletedOnly from '../../components/VisualizationOptions/Options/CompletedOnly'
 import getSeriesTab from './tabs/series'
 import getDataTab from './tabs/data'
-import getAdvancedTemplate from './templates/advanced'
-import getDisplayTemplate from './templates/display'
-import getTotalsTemplate from './templates/totals'
-import getEmptyDataTemplate from './templates/emptyData'
+import getAdvancedTemplate from './sections/templates/advanced'
+import getDisplayTemplate from './sections/templates/display'
+import getTotalsTemplate from './sections/templates/totals'
+import getEmptyDataTemplate from './sections/templates/emptyData'
 import getLegendTab from './tabs/legend'
 import getStyleTab from './tabs/style'
