+++ packages/app/src/modules/options/sections/chartStyle.js
@@ -4,7 +4,7 @@ import React from 'react'
 import HideLegend from '../../../components/VisualizationOptions/Options/HideLegend'
 import NoSpaceBetweenColumns from '../../../components/VisualizationOptions/Options/NoSpaceBetweenColumns'
 import ShowData from '../../../components/VisualizationOptions/Options/ShowData'
-import getChartStyleTemplate from '../templates/chartStyle'
+import getChartStyleTemplate from './templates/chartStyle'
 
 const defaultContent = [<ShowData />, <HideLegend />]
