+++ packages/app/src/modules/options/sections/lines.js
@@ -5,7 +5,7 @@ import i18n from '@dhis2/d2-i18n'
 import RegressionType from '../../../components/VisualizationOptions/Options/RegressionType'
 import TargetLine from '../../../components/VisualizationOptions/Options/TargetLine'
 import BaseLine from '../../../components/VisualizationOptions/Options/BaseLine'
-import getLinesTemplate from '../templates/lines'
+import getLinesTemplate from './templates/lines'
 
 export default hasDisabledSections => ({
     ...getLinesTemplate({
