+++ packages/app/src/modules/options/sections/templates/chartStyle.js
