+++ packages/app/src/modules/options/sections/templates/display.js
