+++ packages/app/src/modules/options/sections/templates/emptyData.js
