+++ packages/app/src/modules/options/sections/templates/totals.js
