+++ packages/app/src/modules/options/sections/templates/verticalAxis.js
